fix: error when serialize the json

+++ shared/src/commonMain/kotlin/com/luckyfrog/quickmart/features/product/domain/entities/ProductEntity.kt
@@ -34,7 +34,7 @@ data class VariantEntity(
     val sku: String?,
     val price: Double?,
     val discountPercentage: Double?,
-    val weight: Int?,
+    val weight: Double?,
     val dimensions: DimensionsEntity?,
     val stock: Int?,
     val size: String?,
